requirements.txt
inmanta-dev-dependencies==1.76.0; python_version <= '3.6'
inmanta-dev-dependencies==2.124.0; python_version > '3.6'
inmanta-core
pyyaml==6.0.1
